Tuuz

+++ app/bot/event/EventMsg.go
@@ -3,16 +3,16 @@ package event
 type EM struct {
 	Type   string `json:"Type"`
 	FromQQ struct {
-		UIN      int64  `json:"UIN"`
+		UIN      int    `json:"UIN"`
 		NickName string `json:"NickName"`
 	} `json:"FromQQ"`
 	OperateQQ struct {
-		UIN      int64  `json:"UIN"`
+		UIN      int    `json:"UIN"`
 		NickName string `json:"NickName"`
 	} `json:"OperateQQ"`
-	LogonQQ   int64 `json:"LogonQQ"`
+	LogonQQ   int `json:"LogonQQ"`
 	FromGroup struct {
-		GIN  int64  `json:"GIN"`
+		GIN  int    `json:"GIN"`
 		Name string `json:"Name"`
 	} `json:"FromGroup"`
 	Msg struct {
